Create PK, move helper into style folder
