+++ flex-config/style/building_combined_point.lua
@@ -1,5 +1,5 @@
 require "helpers"
-require "building_helpers"
+require "style.building_helpers"
 
 
 local tables = {}
